Extract player printing code to a method
Also modified how we validate that the right string is printed. Instead of checking exactness we are checking if the string we are looking is in the total output. Other stuff might also be printed which we're not looking for.

Also first player is the dealer when the game is started.